travel_planner_moya/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
typing_extensions>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
